cleanup(misc): inline assertSupported<Pkg>Version into versions.ts
Moves each per-plugin assertSupported*Version wrapper into the plugin's
versions.ts (next to the minSupported*Version floor it already uses) and removes
the standalone assert-supported-*-version.ts files. No behavior change; the
all-generators-enforce-floor specs still exercise every generator's assert.
