--- packages/detox/src/utils/assert-supported-detox-version.ts
@@ -1,7 +1,0 @@
-import { type Tree } from '@nx/devkit';
-import { assertSupportedPackageVersion } from '@nx/devkit/internal';
-import { minSupportedDetoxVersion } from './versions';
-
-export function assertSupportedDetoxVersion(tree: Tree): void {
-  assertSupportedPackageVersion(tree, 'detox', minSupportedDetoxVersion);
-}
