+++ packages/expo/src/generators/convert-to-inferred/convert-to-inferred.ts
@@ -19,7 +19,7 @@ import { processStartOptions } from './lib/process-start-options';
 import { processSubmitOptions } from './lib/process-submit-options';
 import { processPrebuildOptions } from './lib/process-prebuild-options';
 import { processInstallOptions } from './lib/process-install-options';
-import { assertSupportedExpoVersion } from '../../utils/assert-supported-expo-version';
+import { assertSupportedExpoVersion } from '../../utils/versions';
 
 interface Schema {
   project?: string;
