+++ packages/remix/src/generators/convert-to-inferred/convert-to-inferred.ts
@@ -7,7 +7,7 @@ import {
 import { createNodesV2 } from '../../plugins/plugin';
 import { buildPostTargetTransformer } from './lib/build-post-target-transformer';
 import { servePostTargetTransformer } from './lib/serve-post-target-transformer';
-import { assertSupportedRemixVersion } from '../../utils/assert-supported-remix-version';
+import { assertSupportedRemixVersion } from '../../utils/versions';
 
 interface Schema {
   project?: string;
