+++ packages/remix/src/generators/error-boundary/error-boundary.impl.ts
@@ -1,7 +1,7 @@
 import { formatFiles, type Tree } from '@nx/devkit';
 import { addV2ErrorBoundary, normalizeOptions } from './lib';
 import type { ErrorBoundarySchema } from './schema';
-import { assertSupportedRemixVersion } from '../../utils/assert-supported-remix-version';
+import { assertSupportedRemixVersion } from '../../utils/versions';
 
 export default async function errorBoundaryGenerator(
   tree: Tree,
