+++ packages/remix/src/generators/meta/meta.impl.ts
@@ -1,7 +1,7 @@
 import { Tree } from '@nx/devkit';
 import { v2MetaGenerator } from './lib/v2.impl';
 import { MetaSchema } from './schema';
-import { assertSupportedRemixVersion } from '../../utils/assert-supported-remix-version';
+import { assertSupportedRemixVersion } from '../../utils/versions';
 
 export default async function (tree: Tree, schema: MetaSchema) {
   assertSupportedRemixVersion(tree);
