+++ packages/remix/src/generators/route/route.impl.ts
@@ -7,7 +7,7 @@ import LoaderGenerator from '../loader/loader.impl';
 import MetaGenerator from '../meta/meta.impl';
 import StyleGenerator from '../style/style.impl';
 import { RemixRouteSchema } from './schema';
-import { assertSupportedRemixVersion } from '../../utils/assert-supported-remix-version';
+import { assertSupportedRemixVersion } from '../../utils/versions';
 
 export default async function (tree: Tree, options: RemixRouteSchema) {
   assertSupportedRemixVersion(tree);
